requirements.txt
kafka-python
python-twitter
tweepy
python-dotenv
dash
dash-bootstrap-components
plotly
pandas
pymongo
textblob